requirements.txt
aniso8601==3.0.2
astroid==2.0.4
autopep8==1.3.5
awscli==1.15.39
awsebcli==3.14.1
beautifulsoup4==4.6.0
blessed==1.14.2
boto3==1.7.39
botocore==1.10.39
bs4==0.0.1
cached-property==1.4.2
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
Django==2.0.2
django-cors-headers==2.2.0
django-filter==1.1.0
django-storages==1.6.6
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
docker==3.3.0
docker-compose==1.21.2
docker-pycreds==0.2.3
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
graphene==2.1.3
graphene-django==2.2.0
graphene-file-upload==1.0.0
graphql-core==2.1
graphql-relay==0.4.5
html5lib==1.0.1
idna==2.6
isort==4.3.4
jmespath==0.9.3
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
pathspec==0.5.5
Pillow==5.0.0
promise==2.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.1.1
python-dateutil==2.6.1
python-decouple==3.1
python-resize-image==1.1.11
pytz==2018.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
Rx==1.6.1
s3transfer==0.1.13
semantic-version==2.5.0
singledispatch==3.4.0.3
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.1.0
typing==3.6.6
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.47.0
wrapt==1.10.11
